refactor:split print test code
